requirements.txt
marshmallow~=3.14.0
PyJWT~=2.4.0
Flask~=2.0.2